Change settings path on Windows to CSIDL_APPDATA
For systems where the user profile is loaded from a network drive, the path indicated by CSIDL_LOCAL_APPDATA will only be stored locally and may therefore be lost on logout. CSIDL_APPDATA points to the "roaming" settings folder, which will be accessible on all machines a user can log on to.